RK4 and midpoint working with new solver interface

+++ ParticleToy/particles/Particle.cpp
@@ -12,13 +12,13 @@ void Particle::reset() {
 }
 
 void Particle::draw() {
-    const double h = 0.03;
+    const float h = 0.03;
     glColor3f(1.f, 1.f, 1.f);
     glBegin(GL_QUADS);
-    glVertex2f(m_Position[0] - h / 2.0, m_Position[1] - h / 2.0);
-    glVertex2f(m_Position[0] + h / 2.0, m_Position[1] - h / 2.0);
-    glVertex2f(m_Position[0] + h / 2.0, m_Position[1] + h / 2.0);
-    glVertex2f(m_Position[0] - h / 2.0, m_Position[1] + h / 2.0);
+    glVertex2f(m_Position[0] - h / 2.0f, m_Position[1] - h / 2.0f);
+    glVertex2f(m_Position[0] + h / 2.0f, m_Position[1] - h / 2.0f);
+    glVertex2f(m_Position[0] + h / 2.0f, m_Position[1] + h / 2.0f);
+    glVertex2f(m_Position[0] - h / 2.0f, m_Position[1] + h / 2.0f);
     glEnd();
 }
